% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUNFCCC.R
\name{convertUNFCCC}
\alias{convertUNFCCC}
\title{Convert UNFCCC data}
\usage{
convertUNFCCC(x, subtype = "annex-1")
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readUNFCCC]{readUNFCCC()}}.}

\item{subtype}{Can be either "annex-1" (default) or "non-annex-1"}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Convert UNFCCC data
}
\author{
Falk Benke, Simon Krogmann
}
