% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcommonsenergy-package.R,
%   R/mrenergycore-package.R
\docType{package}
\name{mrcommonsenergy-package}
\alias{mrcommonsenergy-package}
\title{mrcommonsenergy: Fundamental preprocessing functions for REMIND and other energy models}
\description{
Fundamental preprocessing functions for REMIND and other energy models (e.g. EDGE-Transport, EDGE-Buildings).

Fundamental preprocessing functions for REMIND and other energy models (e.g. EDGE-Transport, EDGE-Buildings).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrcommonsenergy}
  \item Report bugs at \url{https://github.com/pik-piam/mrcommonsenergy/issues}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Falk Benke \email{benke@pik-potsdam.de}
}

}
\keyword{internal}
