% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNFCCC.R
\name{readUNFCCC}
\alias{readUNFCCC}
\title{Read UNFCCC data}
\usage{
readUNFCCC(subtype = "annex-1")
}
\arguments{
\item{subtype}{Can be either "annex-1" (default) or "non-annex-1"

filter temporary lock files
finds out if a string is a 4-digit year
finds out if a string is only capital letters A-Z}
}
\description{
Read UNFCCC data
}
\author{
Falk Benke, Simon Krogmann
}
