% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcommonsenergy-package.R
\docType{package}
\name{mrcommonsenergy-package}
\alias{mrcommonsenergy}
\alias{mrcommonsenergy-package}
\title{mrcommonsenergy: Preprocessing functions for REMIND and other energy models (no landuse data)}
\description{
Preprocessing functions for REMIND and other energy models (buildings, transport, industry) not using landuse data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrcommonsenergy}
  \item Report bugs at \url{https://github.com/pik-piam/mrcommonsenergy/issues}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Falk Benke \email{benke@pik-potsdam.de}
  \item Robin Hasse
  \item Renato Rodrigues
  \item Felix Schreyer
}

}
